Collect coverage using API instead of pytest-cov
